Skip controller TLS configuration for controller pod.
The controller proxy connects to the controller via a loopback
interface that is not proxied by the proxy, so the proxy can't add TLS.
TLS isn't needed for loopback interfaces.

Use "localhost." for the override instead of "localhost" since
"localhost" is technically a relative name. Change `conduit install` to
pass through the TLS setting in a way that works for install_test.go.
Modify the existing tests to take into account the above changes.

Add a new install test case that is "the defaults plus TLS."

Signed-off-by: Brian Smith <brian@briansmith.org>